api/,backends/,common/,docs/omsettings,docs/todo.xml, include/om/omdatabase.h,netprogs/,tests/,testsuite/backendmanager.cc, testsuite/backendmanager.h: No longer use OmSettings to specify parameters for constructing databases.  Instead there's a factory function for each database type - temporary naming scheme is OmXxx__open(), mostly because it's easy to grep for later.  At present stub databases and the machinery in InMemory to allow the multierrhandler1 test aren't working.  Everything else should be.
git-svn-id: file:///home/xapian-svn/xapian/trunk@4346 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/backends/quartz/quartz_utils.h
@@ -162,7 +162,7 @@ unpack_uint_last(const char ** src, const char * src_end, T * resultptr)
     // Check byte is what it's meant to be
     CASSERT(sizeof(om_byte) == 1);
 
-    if (src_end - *src > sizeof(T)) {
+    if (src_end - *src > (int)sizeof(T)) {
 	// Would overflow
 	*src = src_end;
 	return false;
